@@ -16,53 +16,72 @@
     "react": "React",
     "react-dom": "ReactDOM"
   },
+  "type": "module",
   "scripts": {
     "build-jsx-browser": "parcel build --log-level 1 --global __root_component --out-dir views views/*.jsx",
     "build-jsx-node": "parcel build --log-level 1 --no-minify --target node --out-dir views/node views/*.jsx",
     "build": "concurrently -i 'npm:build-*'",
-    "start": "node -r esm app.js",
-    "watch-server": "nodemon -e js -r esm app.js",
+    "start": "node --es-module-specifier-resolution=node app.js",
+    "watch-server": "nodemon --es-module-specifier-resolution=node -e js app.js",
     "watch-jsx-browser": "parcel watch --log-level 1 --no-hmr --global __root_component --out-dir views views/*.jsx",
     "watch-jsx-node": "parcel watch --log-level 1 --target node --out-dir views/node views/*.jsx",
     "watch-jsx": "concurrently 'npm:watch-jsx-browser' 'npm:watch-jsx-node'",
     "watch": "concurrently -i 'npm:watch-server' 'npm:watch-jsx-browser' 'npm:watch-jsx-node'",
-    "lint": "eslint app.js server-side-rendering.js routes helpers"
+    "lint": "eslint app.js server-side-rendering.js routes helpers views/**/*.jsx"
   },
   "dependencies": {
-    "@atlaskit/button": "^15.1.0",
-    "@atlaskit/section-message": "^5.0.7",
-    "atlassian-connect-express": "^7.4.7",
-    "atlassian-jwt": "^2.0.0",
-    "axios": "^0.24.0",
-    "body-parser": "^1.19.0",
+    "@atlaskit/button": "^20.2.5",
+    "@atlaskit/section-message": "^6.6.5",
+    "@atlaskit/modal-dialog": "^12.17.6",
+    "@atlaskit/primitives": "^12.2.5",
+    "@atlaskit/form": "^10.5.7",
+    "@atlaskit/textfield": "^6.5.3",
+    "@atlaskit/image": "^1.3.2",
+    "@atlaskit/tokens": "^2.0.3",
+    "@atlaskit/spinner": "^16.3.2",
+    "@onlyoffice/document-editor-react": "^1.5.1",
+    "atlassian-connect-express": "^11.4.1",
+    "atlassian-jwt": "^2.0.3",
+    "axios": "^1.7.7",
+    "body-parser": "^1.20.3",
     "compression": "^1.7.4",
-    "cookie-parser": "^1.4.5",
-    "dotenv": "^16.0.3",
+    "cookie-parser": "^1.4.7",
+    "dotenv": "^16.4.5",
     "errorhandler": "^1.5.1",
     "esm": "^3.2.25",
-    "express": "^4.17.1",
-    "express-hbs": "^2.4.0",
-    "helmet": "^4.1.1",
-    "i18n": "^0.12.0",
+    "express": "^4.21.1",
+    "express-hbs": "^2.5.0",
+    "helmet": "^8.0.0",
+    "i18n": "^0.15.1",
+    "json5": "^2.2.3",
     "morgan": "^1.10.0",
-    "nocache": "^2.1.0",
-    "pg": "^8.10.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "sequelize": "^6.3.5",
-    "styled-components": "^5.2.0"
+    "nocache": "^4.0.0",
+    "pg": "^8.13.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "sequelize": "^6.37.4",
+    "styled-components": "^6.1.13"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "babel-plugin-styled-components": "^1.11.1",
-    "concurrently": "^5.3.0",
-    "eslint": "^7.10.0",
-    "eslint-plugin-react": "^7.21.3",
+    "@babel/core": "^7.25.8",
+    "babel-plugin-styled-components": "^2.1.4",
+    "concurrently": "^9.0.1",
+    "eslint": "^9.12.0",
+    "eslint-plugin-react": "^7.37.1",
     "longjohn": "^0.2.12",
-    "ngrok": "^4.3.0",
-    "nodemon": "^2.0.4",
-    "parcel-bundler": "^1.12.4",
-    "parcel-plugin-externals": "^0.5.1",
-    "sqlite3": "^5.0.0"
+    "ngrok": "^4.3.3",
+    "nodemon": "^3.1.7",
+    "parcel-bundler": "^1.12.5",
+    "parcel-plugin-externals": "^0.5.2",
+    "sqlite3": "^5.1.7"
+  },
+  "overrides": {
+    "braces": "3.0.3",
+    "micromatch": "4.0.8",
+    "node-forge": "1.3.1",
+    "nth-check": "2.1.1",
+    "postcss": "8.4.47",
+    "tough-cookie": "5.0.0",
+    "terser": "4.8.1"
   }
 }
