@@ -46,10 +46,17 @@ import i18n from 'i18n';
 // Routes live here; this is the C in MVC
 import routes from './routes';
 import { addServerSideRendering } from './server-side-rendering';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
+import utils from './helpers/utils';
+import dotenv from 'dotenv';
+
+dotenv.config();
 
 // Bootstrap Express and atlassian-connect-express
 const app = express();
-const addon = ace(app);
+const formats = utils.loadJSON('./resources/assets/document-formats/onlyoffice-docs-formats.json');
+const addon = ace(app, {config: {docServer: {formats: formats}}});
 
 // See config.json
 const port = addon.config.port();
@@ -59,12 +66,15 @@ app.set('port', port);
 const devEnv = app.get('env') === 'development';
 app.use(morgan(devEnv ? 'dev' : 'combined'));
 
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
 // Configure Handlebars
 i18n.configure({
     locales: ['de', 'en', 'es', 'fr', 'it', 'ja', 'ru'],
     defaultLocale: 'en',
     cookie: 'locale',
-    directory: __dirname + "/locales"
+    directory: path.join(__dirname, 'locales')
 });
 
 const viewsDir = path.join(__dirname, 'views');
